Improved loading of RecipeBookConfig and internal way to access it.
