+++ src/main/java/me/wolfyscript/customcrafting/handlers/DataHandler.java
@@ -25,7 +25,6 @@ package me.wolfyscript.customcrafting.handlers;
 import com.google.common.collect.Streams;
 import me.wolfyscript.customcrafting.CustomCrafting;
 import me.wolfyscript.customcrafting.configs.MainConfig;
-import me.wolfyscript.customcrafting.configs.recipebook.Categories;
 import me.wolfyscript.utilities.api.WolfyUtilities;
 import me.wolfyscript.utilities.events.DependenciesLoadedEvent;
 import me.wolfyscript.utilities.util.NamespacedKey;
@@ -47,10 +46,9 @@ public class DataHandler implements Listener {
     public static final File DATA_FOLDER = new File(CustomCrafting.inst().getDataFolder() + File.separator + "data");
     public static final String JSON_OBJ_PATH = DATA_FOLDER.getPath() + "/%s/%s/%s.json";
     private final CustomCrafting customCrafting;
-    private Categories categories;
     private List<Recipe> minecraftRecipes = new ArrayList<>();
 
-    private final MainConfig config;
+    private final ConfigHandler configHandler;
     private final WolfyUtilities api;
     private final DatabaseLoader databaseLoader;
     private final LocalStorageLoader localStorageLoader;
@@ -61,15 +59,16 @@ public class DataHandler implements Listener {
     public DataHandler(CustomCrafting customCrafting) {
         Bukkit.getPluginManager().registerEvents(this, customCrafting);
         this.api = customCrafting.getApi();
-        this.config = customCrafting.getConfigHandler().getConfig();
         this.customCrafting = customCrafting;
+        this.configHandler = customCrafting.getConfigHandler();
 
-        if (customCrafting.getConfigHandler().getConfig().isDatabaseEnabled()) {
+        if (configHandler.getConfig().isDatabaseEnabled()) {
             setSaveDestination(SaveDestination.DATABASE);
             //Currently, there is only support for SQL. MongoDB is planned!
             this.databaseLoader = new SQLDatabaseLoader(customCrafting);
             this.databaseLoader.setPriority(2);
 
+            MainConfig config = configHandler.getConfig();
             if (config.isLocalStorageEnabled()) {
                 this.localStorageLoader = new LocalStorageLoader(customCrafting);
                 this.localStorageLoader.setPriority(config.isLocalStorageBeforeDatabase() ? 3 : 1);
@@ -90,7 +89,6 @@ public class DataHandler implements Listener {
         }
         this.extensionPackLoader = null; //No extension pack implementation yet. TODO
         initLoaders();
-        initCategories();
     }
 
     private void initLoaders() {
@@ -106,11 +104,8 @@ public class DataHandler implements Listener {
         loaders.sort(ResourceLoader::compareTo);
     }
 
-    public void initCategories() {
-        this.categories = customCrafting.getConfigHandler().getRecipeBookConfig().getCategories();
-    }
-
     public void load() {
+        MainConfig config = configHandler.getConfig();
         api.getConsole().info("$msg.startup.recipes.title$");
         var lastBukkitVersion = config.getInt("data.bukkit_version");
         var lastVersion = config.getInt("data.version");
@@ -127,7 +122,7 @@ public class DataHandler implements Listener {
 
     public void loadRecipesAndItems() {
         load();
-        categories.index(customCrafting);
+        configHandler.getRecipeBookConfig().index(customCrafting);
         WorldUtils.getWorldCustomItemStore().initiateMissingBlockEffects();
     }
 
@@ -172,9 +167,5 @@ public class DataHandler implements Listener {
 
     public List<String> getBukkitNamespacedKeys() {
         return getMinecraftRecipes().stream().filter(Keyed.class::isInstance).map(recipe -> NamespacedKey.fromBukkit(((Keyed) recipe).getKey()).toString()).collect(Collectors.toList());
-    }
-
-    public Categories getCategories() {
-        return categories;
     }
 }
