+++ src/main/java/me/wolfyscript/customcrafting/utils/PlayerUtil.java
@@ -51,7 +51,7 @@ public class PlayerUtil {
     public static void openRecipeBook(Player player) {
         CustomCrafting customCrafting = CustomCrafting.inst();
         InventoryAPI<CCCache> invAPI = customCrafting.getApi().getInventoryAPI(CCCache.class);
-        var categories = customCrafting.getDataHandler().getCategories();
+        var categories = customCrafting.getConfigHandler().getRecipeBookConfig();
 
         // Open directly to the category if we only have one
         if (categories.getSortedCategories().size() == 1) {
